Update LogRecord to 64bit event version.

+++ src/EventStore.Core/Services/Storage/ReaderIndex/IndexCommitter.cs
@@ -363,8 +363,7 @@ namespace EventStore.Core.Services.Storage.ReaderIndex
                 var entries = _tableIndex.GetRange(streamId, indexEntries[0].Version, indexEntries[indexEntries.Count - 1].Version);
                 foreach (var indexEntry in entries)
                 {
-                    // TODO HAYLEY: Change this back to a long
-                    var prepare = prepares[(int)indexEntry.Version - (int)indexEntries[0].Version];
+                    var prepare = prepares[indexEntry.Version - indexEntries[0].Version];
                     PrepareLogRecord indexedPrepare = GetPrepare(reader, indexEntry.Position);
                     if (indexedPrepare != null && indexedPrepare.EventStreamId == prepare.EventStreamId)
                     {
